update per pr feedback
